+++ Cargo.toml
@@ -21,8 +21,9 @@ iced = { version = "0.13.1", features = [
     "lazy",
     "advanced",
 ] }
-tokio = { version = "1.36", features = ["full"] }
+tokio = { version = "1.40.0", features = ["full"] }
 iced_graphics = "0.13.0"
 iced_core = "0.13.2"
 open = "5.3.0"
 anyhow = "1.0.89"
+futures = "0.3.30"
