+++ src/settings.rs
@@ -1,0 +1,53 @@
+use iced::Font;
+
+use crate::ColorPalette;
+
+const DEFAULT_SHELL: &str = "/bin/bash";
+
+#[derive(Default, Clone)]
+pub struct Settings {
+    pub font: FontSettings,
+    pub theme: ThemeSettings,
+    pub backend: BackendSettings,
+}
+
+#[derive(Debug, Clone)]
+pub struct BackendSettings {
+    pub shell: String,
+}
+
+impl Default for BackendSettings {
+    fn default() -> Self {
+        Self {
+            shell: DEFAULT_SHELL.to_string(),
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct FontSettings {
+    pub size: f32,
+    pub scale_factor: f32,
+    pub font_type: Font,
+}
+
+impl Default for FontSettings {
+    fn default() -> Self {
+        Self {
+            size: 14.0,
+            scale_factor: 1.3,
+            font_type: Font::MONOSPACE,
+        }
+    }
+}
+
+#[derive(Default, Debug, Clone)]
+pub struct ThemeSettings {
+    pub color_pallete: Box<ColorPalette>
+}
+
+impl ThemeSettings {
+    pub fn new(color_pallete: Box<ColorPalette>) -> Self {
+        Self { color_pallete }
+    }
+}
